fix for 7, test

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -4,7 +4,7 @@
 * @n: automatic int var to take input
 * Return: last digit of the number
 */
-int print_last_digit(long int n)
+int print_last_digit(int n)
 {
 int last;
 if (n < 0)
